feat: The share button is added to the log page
